Models, Forms ,Styles, migrations and seeders setup.
Set up, forms and styled them as needed.
Set up some seeders, currently using only the rolesSeeder. The others are going to be needed in the future.
Fixed some routes in web.php that were not correct.
Created migrations files for all forms.
Edited the models where i had to set the tables & relations. (WIP) .
Added loading accidents along with all the data in the AccidentController.
In UserController, on user registration the default role_id is set to 1 which represents the "Registrar" role.(Contained inside the rolesSeeder.php file).